requirements.txt
# ArduPilot Drone AI - Requirements

# MAVLink communication
pymavlink>=2.4.41

# Async support
aiohttp>=3.8.0

# Configuration
pyyaml>=6.0

# Logging
structlog>=23.1.0

# Numerical computing
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
